refactor: move fetchAvailableVersions to shared-internal
