+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/updating/GitHubPullRequestCreation.kt
@@ -1,6 +1,5 @@
 package io.github.typesafegithub.workflows.codegenerator.updating
 
-import io.github.typesafegithub.workflows.codegenerator.versions.httpClient
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.HttpRequestBuilder
