+++ maven-binding-builder/build.gradle.kts
@@ -1,16 +1,11 @@
 plugins {
     buildsrc.convention.`kotlin-jvm`
-    kotlin("plugin.serialization")
 }
 
 dependencies {
     implementation("org.jetbrains.kotlin:kotlin-compiler")
     api(projects.actionBindingGenerator)
-    implementation(platform("io.ktor:ktor-bom:2.3.10"))
-    implementation("io.ktor:ktor-client-core")
-    implementation("io.ktor:ktor-client-cio")
-    implementation("io.ktor:ktor-client-content-negotiation")
-    implementation("io.ktor:ktor-serialization-kotlinx-json")
+    implementation(projects.sharedInternal)
 
     runtimeOnly(projects.githubWorkflowsKt)
 }
