+++ shared-internal/src/main/kotlin/io/github/typesafegithub/workflows/shared/internal/GithubApi.kt
@@ -1,0 +1,63 @@
+package io.github.typesafegithub.workflows.shared.internal
+
+import io.github.typesafegithub.workflows.shared.internal.model.Version
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
+import io.ktor.client.request.bearerAuth
+import io.ktor.client.request.get
+import io.ktor.serialization.kotlinx.json.json
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+
+suspend fun fetchAvailableVersions(
+    owner: String,
+    name: String,
+    githubToken: String,
+): List<Version> =
+    listOf(
+        apiTagsUrl(owner = owner, name = name),
+        apiBranchesUrl(owner = owner, name = name),
+    ).flatMap { url -> fetchGithubRefs(url, githubToken) }
+        .versions()
+
+private fun List<GithubRef>.versions(): List<Version> =
+    this.map { githubRef ->
+        val version = githubRef.ref.substringAfterLast("/")
+        Version(version)
+    }
+
+private suspend fun fetchGithubRefs(
+    url: String,
+    githubToken: String,
+): List<GithubRef> =
+    httpClient.get(urlString = url) {
+        bearerAuth(githubToken)
+    }.body()
+
+private fun apiTagsUrl(
+    owner: String,
+    name: String,
+): String = "https://api.github.com/repos/$owner/$name/git/matching-refs/tags/v"
+
+private fun apiBranchesUrl(
+    owner: String,
+    name: String,
+): String = "https://api.github.com/repos/$owner/$name/git/matching-refs/heads/v"
+
+@Serializable
+private data class GithubRef(
+    val ref: String,
+)
+
+private val httpClient by lazy {
+    HttpClient {
+        install(ContentNegotiation) {
+            json(
+                Json {
+                    ignoreUnknownKeys = true
+                },
+            )
+        }
+    }
+}
